pyproject.toml
[project]
name = "test-case-sender"
version = "0.1.0"
description = ""
authors = [
    {name = "fedor",email = "fedorlastv@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "taskiq (>=0.11.17,<0.12.0)",
    "dishka (>=1.6.0,<2.0.0)",
    "taskiq-aio-pika (>=0.4.2,<0.5.0)",
    "opentelemetry-api (>=1.34.0,<2.0.0)",
    "opentelemetry-sdk (>=1.34.0,<2.0.0)",
    "opentelemetry-exporter-prometheus (>=0.55b0,<0.56)",
    "prometheus-client (>=0.22.1,<0.23.0)",
    "opentelemetry-exporter-otlp (>=1.34.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
